Add possibility to have 2 site titles separated with `|`
Set `title` as an optional config parameter and use `site_title` as the main one.
If param `title` exists, site title will be displayed as "site_title | title"
otherwise, it will be just "site_title".

It has a fallback strategy for when "site_title" is not configured -> then it just takes the title from index.html